requirements.txt
httpx>=0.27,<1
pydantic>=2.7,<3
PyYAML>=6.0,<7

